备份

+++ easy-framework-core/src/main/java/com/yffd/easy/framework/core/common/dao/BaseDaoImpl.java
@@ -88,7 +88,7 @@ public class BaseDaoImpl<T> extends EasyModelConverter implements IBaseDao<T> {
         }
         paramMap.remove("pageParam"); // 与 listRange区分，保证越过分页拦截器
         // 统计总记录数
-        Long totalRecord = sqlSession.selectOne(getStatement(SQL_SELECT_COUNT_BY), paramMap);
+        int totalRecord = sqlSession.selectOne(getStatement(SQL_SELECT_COUNT_BY), paramMap);
         pageParam.setTotalRecord(totalRecord);
         paramMap.put("pageParam", pageParam);
         // 获取分页数据集
